requirements.txt
langchain>=1.0.0
langchain-core>=1.0.0
langchain-community>=0.2.0
langchain-openai>=0.1.0
langgraph>=1.0.0
python-dotenv>=1.0.0
openai>=1.40.0

chromadb>=0.5.0
requests>=2.32.0
pydantic>=2.10.0
typing-extensions>=4.0.0
langgraph-checkpoint>=1.0.0
langgraph-checkpoint-sqlite>=2.0.11
streamlit>=1.50.0
yfinance>=0.2.0